+++ packages/fighting-design/image-preview/src/image-preview.vue
@@ -1,0 +1,12 @@
+<script lang="ts" setup name="FImagePreview">
+  import { Props, Emits } from './image-preview'
+
+  defineProps(Props)
+  defineEmits(Emits)
+</script>
+
+<template>
+  <div class="f-image-preview">
+    FImagePreview
+  </div>
+</template>